Replication/requirements.txt
anthropic==0.52.1
javalang==0.13.0
numpy==1.26.4
ollama==0.5.1
openai==1.82.1
pandas==2.2.3
psutil==7.0.0
PyYAML==6.0.2
scikit_learn==1.6.1
tiktoken==0.9.0
torch==2.7.0
transformers==4.52.4
